fix: split the collector into collector+extractor
