+++ backend/plugins/github_graphql/impl/impl.go
@@ -81,53 +81,56 @@ func (p GithubGraphql) GetTablesInfo() []dal.Tabler {
 
 func (p GithubGraphql) SubTaskMetas() []plugin.SubTaskMeta {
 	return []plugin.SubTaskMeta{
-		//tasks.CollectRepoMeta,
-
 		// collect millstones
 		githubTasks.CollectMilestonesMeta,
 		githubTasks.ExtractMilestonesMeta,
 
 		// collect issue & pr, deps on millstone
-		tasks.CollectIssueMeta,
-		tasks.CollectPrMeta,
+		tasks.CollectIssuesMeta,
+		tasks.ExtractIssuesMeta,
+		tasks.CollectPrsMeta,
+		tasks.ExtractPrsMeta,
 
 		// collect workflow run & job
 		githubTasks.CollectRunsMeta,
 		githubTasks.ExtractRunsMeta,
-		tasks.CollectGraphqlJobsMeta,
+		tasks.CollectJobsMeta,
+		tasks.ExtractJobsMeta,
 
-		// collect others
-		githubTasks.CollectApiCommentsMeta,
-		githubTasks.ExtractApiCommentsMeta,
-		githubTasks.CollectApiEventsMeta,
-		githubTasks.ExtractApiEventsMeta,
-		githubTasks.CollectApiPrReviewCommentsMeta,
-		githubTasks.ExtractApiPrReviewCommentsMeta,
+		// // collect others
+		// githubTasks.CollectApiCommentsMeta,
+		// githubTasks.ExtractApiCommentsMeta,
+		// githubTasks.CollectApiEventsMeta,
+		// githubTasks.ExtractApiEventsMeta,
+		// githubTasks.CollectApiPrReviewCommentsMeta,
+		// githubTasks.ExtractApiPrReviewCommentsMeta,
 
-		// collect account, deps on all before
-		tasks.CollectAccountMeta,
+		// // collect account, deps on all before
+		// tasks.CollectAccountMeta,
+		// tasks.ExtractAccountsMeta,
 
-		// convert to domain layer
-		githubTasks.ConvertRunsMeta,
-		githubTasks.ConvertJobsMeta,
-		githubTasks.EnrichPullRequestIssuesMeta,
-		githubTasks.ConvertRepoMeta,
-		githubTasks.ConvertIssuesMeta,
-		githubTasks.ConvertCommitsMeta,
-		githubTasks.ConvertIssueLabelsMeta,
-		githubTasks.ConvertPullRequestCommitsMeta,
-		githubTasks.ConvertPullRequestsMeta,
-		githubTasks.ConvertPullRequestReviewsMeta,
-		githubTasks.ConvertPullRequestLabelsMeta,
-		githubTasks.ConvertPullRequestIssuesMeta,
-		githubTasks.ConvertIssueAssigneeMeta,
-		githubTasks.ConvertIssueCommentsMeta,
-		githubTasks.ConvertPullRequestCommentsMeta,
-		githubTasks.ConvertMilestonesMeta,
-		githubTasks.ConvertAccountsMeta,
+		// // convert to domain layer
+		// githubTasks.ConvertRunsMeta,
+		// githubTasks.ConvertJobsMeta,
+		// githubTasks.EnrichPullRequestIssuesMeta,
+		// githubTasks.ConvertRepoMeta,
+		// githubTasks.ConvertIssuesMeta,
+		// githubTasks.ConvertCommitsMeta,
+		// githubTasks.ConvertIssueLabelsMeta,
+		// githubTasks.ConvertPullRequestCommitsMeta,
+		// githubTasks.ConvertPullRequestsMeta,
+		// githubTasks.ConvertPullRequestReviewsMeta,
+		// githubTasks.ConvertPullRequestLabelsMeta,
+		// githubTasks.ConvertPullRequestIssuesMeta,
+		// githubTasks.ConvertIssueAssigneeMeta,
+		// githubTasks.ConvertIssueCommentsMeta,
+		// githubTasks.ConvertPullRequestCommentsMeta,
+		// githubTasks.ConvertMilestonesMeta,
+		// githubTasks.ConvertAccountsMeta,
 
 		// deployment
-		tasks.CollectAndExtractDeploymentsMeta,
+		tasks.CollectDeploymentsMeta,
+		tasks.ExtractDeploymentsMeta,
 		githubTasks.ConvertDeploymentsMeta,
 	}
 }
