+++ backend/plugins/github_graphql/tasks/issue_collector.go
@@ -21,12 +21,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/apache/incubator-devlake/core/dal"
 	"github.com/apache/incubator-devlake/core/errors"
-	"github.com/apache/incubator-devlake/core/models/domainlayer/ticket"
 	"github.com/apache/incubator-devlake/core/plugin"
 	helper "github.com/apache/incubator-devlake/helpers/pluginhelper/api"
-	"github.com/apache/incubator-devlake/plugins/github/models"
 	githubTasks "github.com/apache/incubator-devlake/plugins/github/tasks"
 	"github.com/merico-dev/graphql"
 )
@@ -73,30 +70,18 @@ type GraphqlQueryIssue struct {
 	} `graphql:"labels(first: 100)"`
 }
 
-var CollectIssueMeta = plugin.SubTaskMeta{
-	Name:             "CollectIssue",
-	EntryPoint:       CollectIssue,
+var CollectIssuesMeta = plugin.SubTaskMeta{
+	Name:             "CollectIssues",
+	EntryPoint:       CollectIssues,
 	EnabledByDefault: true,
 	Description:      "Collect Issue data from GithubGraphql api, supports both timeFilter and diffSync.",
 	DomainTypes:      []string{plugin.DOMAIN_TYPE_TICKET},
 }
 
-var _ plugin.SubTaskEntryPoint = CollectIssue
+var _ plugin.SubTaskEntryPoint = CollectIssues
 
-func CollectIssue(taskCtx plugin.SubTaskContext) errors.Error {
-	db := taskCtx.GetDal()
+func CollectIssues(taskCtx plugin.SubTaskContext) errors.Error {
 	data := taskCtx.GetData().(*githubTasks.GithubTaskData)
-	config := data.Options.ScopeConfig
-	issueRegexes, err := githubTasks.NewIssueRegexes(config)
-	if err != nil {
-		return nil
-	}
-
-	milestoneMap, err := getMilestoneMap(db, data.Options.GithubId, data.Options.ConnectionId)
-	if err != nil {
-		return nil
-	}
-
 	collectorWithState, err := helper.NewStatefulApiCollector(helper.RawDataSubTaskArgs{
 		Ctx: taskCtx,
 		Params: githubTasks.GithubApiParams{
@@ -136,42 +121,7 @@ func CollectIssue(taskCtx plugin.SubTaskContext) errors.Error {
 			return query.Repository.IssueList.PageInfo, nil
 		},
 		ResponseParser: func(iQuery interface{}, variables map[string]interface{}) ([]interface{}, error) {
-			query := iQuery.(*GraphqlQueryIssueWrapper)
-			issues := query.Repository.IssueList.Issues
-
-			results := make([]interface{}, 0, 1)
-			isFinish := false
-			for _, issue := range issues {
-				githubIssue, err := convertGithubIssue(milestoneMap, issue, data.Options.ConnectionId, data.Options.GithubId)
-				if err != nil {
-					return nil, err
-				}
-				githubLabels, err := convertGithubLabels(issueRegexes, issue, githubIssue)
-				if err != nil {
-					return nil, err
-				}
-				results = append(results, githubLabels...)
-				results = append(results, githubIssue)
-				if len(issue.AssigneeList.Assignees) > 0 {
-					extractGraphqlPreAccount(&results, &issue.AssigneeList.Assignees[0], data.Options.GithubId, data.Options.ConnectionId)
-				}
-				extractGraphqlPreAccount(&results, issue.Author, data.Options.GithubId, data.Options.ConnectionId)
-				for _, assignee := range issue.AssigneeList.Assignees {
-					issueAssignee := &models.GithubIssueAssignee{
-						ConnectionId: githubIssue.ConnectionId,
-						IssueId:      githubIssue.GithubId,
-						RepoId:       githubIssue.RepoId,
-						AssigneeId:   assignee.Id,
-						AssigneeName: assignee.Login,
-					}
-					results = append(results, issueAssignee)
-				}
-			}
-			if isFinish {
-				return results, helper.ErrFinishCollect
-			} else {
-				return results, nil
-			}
+			return nil, nil
 		},
 	})
 	if err != nil {
@@ -179,102 +129,4 @@ func CollectIssue(taskCtx plugin.SubTaskContext) errors.Error {
 	}
 
 	return collectorWithState.Execute()
-}
-
-// create a milestone map for numberId to databaseId
-func getMilestoneMap(db dal.Dal, repoId int, connectionId uint64) (map[int]int, errors.Error) {
-	milestoneMap := map[int]int{}
-	var milestones []struct {
-		MilestoneId int
-		RepoId      int
-		Number      int
-	}
-	err := db.All(
-		&milestones,
-		dal.From(&models.GithubMilestone{}),
-		dal.Where("repo_id = ? and connection_id = ?", repoId, connectionId),
-	)
-	if err != nil {
-		return nil, err
-	}
-	for _, milestone := range milestones {
-		milestoneMap[milestone.Number] = milestone.MilestoneId
-	}
-	return milestoneMap, nil
-}
-
-func convertGithubIssue(milestoneMap map[int]int, issue GraphqlQueryIssue, connectionId uint64, repositoryId int) (*models.GithubIssue, errors.Error) {
-	githubIssue := &models.GithubIssue{
-		ConnectionId:    connectionId,
-		GithubId:        issue.DatabaseId,
-		RepoId:          repositoryId,
-		Number:          issue.Number,
-		State:           issue.State,
-		Title:           issue.Title,
-		Body:            strings.ReplaceAll(issue.Body, "\x00", `<0x00>`),
-		Url:             issue.Url,
-		ClosedAt:        issue.ClosedAt,
-		GithubCreatedAt: issue.CreatedAt,
-		GithubUpdatedAt: issue.UpdatedAt,
-	}
-	if issue.AssigneeList.Assignees != nil && len(issue.AssigneeList.Assignees) > 0 {
-		githubIssue.AssigneeId = issue.AssigneeList.Assignees[0].Id
-		githubIssue.AssigneeName = issue.AssigneeList.Assignees[0].Login
-	}
-	if issue.Author != nil {
-		githubIssue.AuthorId = issue.Author.Id
-		githubIssue.AuthorName = issue.Author.Login
-	}
-	if issue.ClosedAt != nil {
-		githubIssue.LeadTimeMinutes = uint(issue.ClosedAt.Sub(issue.CreatedAt).Minutes())
-	}
-	if issue.Milestone != nil {
-		if milestoneId, ok := milestoneMap[issue.Milestone.Number]; ok {
-			githubIssue.MilestoneId = milestoneId
-		}
-	}
-	return githubIssue, nil
-}
-
-func convertGithubLabels(issueRegexes *githubTasks.IssueRegexes, issue GraphqlQueryIssue, githubIssue *models.GithubIssue) ([]interface{}, errors.Error) {
-	var results []interface{}
-	var joinedLabels []string
-	for _, label := range issue.Labels.Nodes {
-		results = append(results, &models.GithubIssueLabel{
-			ConnectionId: githubIssue.ConnectionId,
-			IssueId:      githubIssue.GithubId,
-			LabelName:    label.Name,
-		})
-		joinedLabels = append(joinedLabels, label.Name)
-
-		if issueRegexes.SeverityRegex != nil {
-			groups := issueRegexes.SeverityRegex.FindStringSubmatch(label.Name)
-			if len(groups) > 1 {
-				githubIssue.Severity = groups[1]
-			}
-		}
-		if issueRegexes.ComponentRegex != nil {
-			groups := issueRegexes.ComponentRegex.FindStringSubmatch(label.Name)
-			if len(groups) > 1 {
-				githubIssue.Component = groups[1]
-			}
-		}
-		if issueRegexes.PriorityRegex != nil {
-			groups := issueRegexes.PriorityRegex.FindStringSubmatch(label.Name)
-			if len(groups) > 1 {
-				githubIssue.Priority = groups[1]
-			}
-		}
-		if issueRegexes.TypeRequirementRegex != nil && issueRegexes.TypeRequirementRegex.MatchString(label.Name) {
-			githubIssue.StdType = ticket.REQUIREMENT
-		} else if issueRegexes.TypeBugRegex != nil && issueRegexes.TypeBugRegex.MatchString(label.Name) {
-			githubIssue.StdType = ticket.BUG
-		} else if issueRegexes.TypeIncidentRegex != nil && issueRegexes.TypeIncidentRegex.MatchString(label.Name) {
-			githubIssue.StdType = ticket.INCIDENT
-		}
-	}
-	if len(joinedLabels) > 0 {
-		githubIssue.Type = strings.Join(joinedLabels, ",")
-	}
-	return results, nil
 }
